pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-workbench"
version = "0.1.0"
description = "Production-minded AI agent workbench with iterative workflows"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "joblib>=1.3.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "prometheus-client>=0.19.0",
    "click>=8.1.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "readability-lxml>=0.8.1",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "chromadb>=0.4.0",
    "trafilatura>=0.10.0",
  "lxml-html-clean>=0.2.0",
  "jsonschema>=4.19.0",
  "pluggy>=1.3.0",
  "networkx>=3.2.1",
  "orjson>=3.9.15",
]

[project.scripts]
aw = "agent_workbench.cli:main"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]
